rvc: rename local package ttnn/ -> tt/ to match tt-metal audio convention
The local RVC TTNN implementation lived in models/demos/rvc/ttnn/, which
collides with the system ``ttnn`` Python package whenever pytest's default
import-mode adds models/demos/rvc/ to sys.path. The collision manifested
as ``AttributeError: partially initialized module 'ttnn' has no attribute
'DRAM_MEMORY_CONFIG' (most likely due to a circular import)`` and required
every reviewer to pass ``--import-mode=importlib`` to pytest as a workaround.

This rename eliminates the shadowing at the package layer:

  models/demos/rvc/ttnn/  -> models/demos/rvc/tt/

The new name matches the tt-metal audio model convention
(``models/demos/audio/whisper/tt/``).

Changes:
- ``git mv`` the entire directory.
- Update 8 import sites across demo.py, benchmark.py, profile.py, the
  test files, and the package's own __init__.py / runtime.py / ops files.
- Refresh the module docstring in tt/__init__.py to reflect the new name
  and note the convention.

Verification on N300:
- ``pytest models/demos/rvc/tests/test_runtime.py -q``           5/5 PASS
- ``pytest models/demos/rvc/tests/test_production_shapes.py -q`` 9/9 PASS
- ``pytest models/demos/rvc/tests/test_ttnn_ops.py -q``         39/39 PASS
  All three now pass without --import-mode=importlib.
- Full demo end-to-end: Flow PCC 0.999986, Audio PCC 0.997669.
- benchmark.py warm 3 s: TTNN-only RTF 0.539, full RTF 0.665, PCC 0.998.

No inference-path code semantics changed.
